用户页面的查询

+++ question-service/src/main/java/com/paper/question/service/SysUserImpl.java
@@ -3,8 +3,6 @@ package com.paper.question.service;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.paper.question.common.PageResult;
-import com.paper.question.common.Pagination;
-import com.paper.question.dao.mapper.SysRoleMapper;
 import com.paper.question.dao.mapper.SysUserMapper;
 import com.paper.question.dao.mapper.SysUserRoleMapper;
 import com.paper.question.domain.dto.SysUserDto;
@@ -27,10 +25,12 @@ public class SysUserImpl implements ISysUserService{
     private SysUserRoleMapper sysUserRoleMapper;
 
     @Override
-    public PageResult list(Pagination pagination) {
+    public PageResult list( SysUser pagination) {
+        System.out.println("用户分页");
+        System.out.println(pagination);
         Integer pageNum = PageResult.PageInfo.getPageNum(pagination.getPageNum());
         Integer pageSize = PageResult.PageInfo.getPageSize(pagination.getPageSize());
-        Page<SysUserDto> page = PageHelper.startPage(pageNum, pageSize).doSelectPage(() -> sysUserMapper.list());
+        Page<SysUserDto> page = PageHelper.startPage(pageNum, pageSize).doSelectPage(() -> sysUserMapper.list(pagination));
         return new PageResult().get(page);
     }
 
@@ -55,18 +55,33 @@ public class SysUserImpl implements ISysUserService{
 
     @Override
     public int editUser(SysUserEditDto sysUserEditDto) {
-        //更新用户的信息
-       int userId = sysUserMapper.updateByPrimaryKeySelective(sysUserEditDto.getSysUser());
-       //先将用户的角色进行删除再进行添加
+        int userId = 0;
+        if(sysUserEditDto.getSysUser().getId() != null){
+            //更新用户的信息
+            userId =  sysUserMapper.updateByPrimaryKeySelective(sysUserEditDto.getSysUser());
+
+        }else{
+            //初始密码为123456
+           sysUserEditDto.getSysUser().setPassword("123456");
+           userId = sysUserMapper.insertSelective(sysUserEditDto.getSysUser());
+        }
+        //先将用户的角色进行删除再进行添加
         List<Long> roleId = sysUserEditDto.getRoleIds();
-        roleId.forEach(item -> {
-            sysUserRoleMapper.deleteByPrimaryKey(item);
-            SysUserRole sysUserRole = new SysUserRole();
-            sysUserRole.setRoleId(item);
-            long uid = userId;
-            sysUserRole.setUserId(uid);
-            sysUserRoleMapper.insertSelective(sysUserRole);
-         });
+        System.out.println("角色信息");
+        System.out.println(roleId);
+        if(roleId != null){
+            for (Long item: roleId) {
+                sysUserRoleMapper.deleteByPrimaryKey(item);
+                SysUserRole sysUserRole = new SysUserRole();
+                sysUserRole.setRoleId(item);
+                long uid = userId;
+                sysUserRole.setUserId(uid);
+                sysUserRoleMapper.insertSelective(sysUserRole);
+            }
+        }
+//        roleId.forEach(item -> {
+//
+//        });
 
         return userId;
     }
@@ -74,5 +89,18 @@ public class SysUserImpl implements ISysUserService{
     @Override
     public int deleteUser(long userId) {
         return sysUserMapper.deleteByPrimaryKey(userId);
+    }
+
+    @Override
+    public int batchDelete(Long[] userIds) {
+        return sysUserMapper.batchDelete(userIds);
+    }
+
+    @Override
+    public int changeStatus(long id, Integer status) {
+        SysUser sysUser = new SysUser();
+        sysUser.setId(id);
+        sysUser.setStatus(status);
+        return sysUserMapper.updateByPrimaryKeySelective(sysUser);
     }
 }
